Add abstraction level in controllers and fix adding dish
Whole tree of abstract controllers classes added to unify controllers communication.
Functionality of loading controllers and their scenes was moved to new utility class
"Controllers" which is more readable and provides easy way to add new scenes and
controllers to application. FXMLLoader was wrapped into "FXMLSceneLoader" class
to catch checked exceptions which may occure and deal with them on low level of app.